fix: change camera preview to full screen size
